Refactor data model and optimize filtering.
Changed primary keys from numeric to bigint in several tables to improve data consistency and future-proof the schema. Added null checks to ensure data integrity during map operations. Removed unnecessary imports in AssertionDialog.

+++ src/main/java/io/github/mtrevisan/familylegacy/flef/persistence/models/ResearchStatusEntity.java
@@ -73,7 +73,7 @@ public class ResearchStatusEntity extends AbstractEntity{
 	private String status;
 
 	@Column(name = "priority")
-	private Integer priority;
+	private Short priority;
 
 	@Column(name = "creation_date", columnDefinition= "TIMESTAMP WITH TIME ZONE", nullable = false)
 	private ZonedDateTime creationDate;
